# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Xpart App OpenApi API Documentation",
        "description": "Xpart App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/shared/withdrawals": {
            "post": {
                "tags": [
                    "Shared"
                ],
                "summary": "postWithdrawals",
                "description": "postWithdrawals",
                "operationId": "postWithdrawals",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WithdrawalCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "WithdrawalCreateFormRequest": {
                "title": "Withdrawal Create Form Request Fields",
                "description": "Withdrawal Create Form request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "amount": {
                        "title": "Withdrawal amount",
                        "description": "Withdrawal amount ",
                        "type": "integer",
                        "example": "1000"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}